Incorporated red-black binary tree function family, changed ft_putchar to align it with POSIX and ISO C standard, changed ft_putnbr_fd to accept long long int, added defines for ANSI color codes, added ft_get_window_size, moved header files to includes directory

+++ ft_putnbr_fd.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mrosario <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: miki <miki@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/13 13:36:08 by mrosario          #+#    #+#             */
-/*   Updated: 2019/11/13 18:18:09 by mrosario         ###   ########.fr       */
+/*   Updated: 2021/04/07 00:51:40 by miki             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(long long int n, int fd)
 {
 	long long int	ncpy;
